Tests/LibWeb/Text/expected/canvas/textmetrics-actual-bounding-box-accent.txt

precomposed accent has positive width: true
precomposed and decomposed accent widths match: true
word width includes precomposed accent: true
